+++ R/is.xifti.R
@@ -91,7 +91,7 @@ is.xifti_data <- function(x) {
 #' @return Logical indicating whether x is a valid surface.
 #' @export
 #' 
-is.surface <- function(x) {
+is.surf <- function(x) {
   if (!match_exactly(names(x), c("vertices", "faces"))) {
     return(FALSE)
   }
@@ -375,7 +375,7 @@ is.xifti <- function(x, messages=TRUE) {
 
   if (!is.xifti_data(x$data)) { message('"data" is invalid.\n'); return(FALSE) }
   for (s in names(x$surf)) {
-    if (!is.null(x$surf[[s]]) && !is.surface(x$surf[[s]])) {
+    if (!is.null(x$surf[[s]]) && !is.surf(x$surf[[s]])) {
       message(paste(s, "is invalid.\n")); return(FALSE)
     }
   }