+++ R/utils_volume.R
@@ -15,7 +15,7 @@
 #'
 #' @export
 #' 
-unmask <- function(dat, mask, fill=NA) {
+unmask_vol <- function(dat, mask, fill=NA) {
 
   # Check that dat is a vector or matrix.
   if (is.vector(dat) || is.factor(dat)) { dat <- matrix(dat, ncol=1) }
@@ -49,7 +49,7 @@ unmask <- function(dat, mask, fill=NA) {
 #' Pad a 3D array by a certain amount in each direction, along each dimension.
 #'  This effectively undoes a crop.
 #'
-#' @param x A 3D array, e.g. \code{unmask(xifti$data$subcort, xifti$meta$subcort$mask)}.
+#' @param x A 3D array, e.g. \code{unmask_vol(xifti$data$subcort, xifti$meta$subcort$mask)}.
 #' @param padding A dx2 matrix indicating the number of 
 #'  slices to add at the beginning and end of each of the d dimensions, e.g.
 #'  \code{xifti$meta$subcort$mask_padding}.