+++ R/write_nifti.R
@@ -20,12 +20,12 @@ write_subcort_nifti <- function(
   subcortVol_fname, subcortLabs_fname, fill=0, wb_path=NULL){
 
   # Data.
-  writeNifti(unmask(subcortVol, subcortMask, fill=fill), subcortVol_fname)
+  writeNifti(unmask_vol(subcortVol, subcortMask, fill=fill), subcortVol_fname)
 
   # Labels...
   stopifnot(is.subcort_labs(subcortLabs))
   subcortLabs <- as.numeric(subcortLabs) #- 2
-  writeNifti(unmask(subcortLabs, subcortMask, fill=fill), subcortLabs_fname)
+  writeNifti(unmask_vol(subcortLabs, subcortMask, fill=fill), subcortLabs_fname)
   
   # ...Add back subcortical label information.
   # https://www.humanconnectome.org/software/workbench-command/-volume-help